#230 chore: refactor add model service layer
